Tests, if fires up the right part of component
[What?]
The rendering component is dependent of URL.
So, here I'm testing, if the labels are in relation with URL.

[Why?]
To be sure, that f.e. when register URL, the label is "Create an account"

[How?]
So, if you want to test the rendering part of component, I need to set the URL to
the right value.
Then check if labels are as expected.